added `dist` to the repository and `public` to the .gitignore
`dist` will be used as the consumable part of a plugin/library. a user that is including the plugin will use the compiled versions using the `dist` folder. this will let each user have access to the post-compiled version of the code, and not require them to install/learn/setup compilers like DART or typescript.